requirements.txt
six==1.16.
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
flask-talisman==1.1.0
Flask-Login==0.6.2
flask-compress==1.14
google-cloud-ndb==2.2.0
google-cloud-tasks==2.14.1
google-cloud-storage==2.12.0
google-cloud-vision==3.4.5
google_cloud_documentai==2.20.0
coolname==2.2.0
openai==1.3.5
psutil==5.9.5
featurebase==0.0.2
openai[datalib]
confluent-kafka==2.2.0
markdown==3.4.4
pytest==7.4.1
pytest-cov==4.1.0
ping3==4.0.4
apscheduler==3.10.4
Faker==19.11.0
PyPDF2==3.0.1
nltk==3.8.1
phonenumbers==8.13.25
email_validator==2.1.0
twilio==8.10.2
sendgrid==6.10.0
slackclient==2.9.4
Pillow==10.1.0
pydub==0.25.1
ffmpeg-python==0.2.0